Commit -WIP sass files
- Adding new sass directories and refators
