+++ package.json
@@ -29,7 +29,6 @@
     "eslint-plugin-react-hooks": "^5.1.0-rc.0",
     "eslint-plugin-react-refresh": "^0.4.12",
     "globals": "^15.9.0",
-    "sass": "^1.80.4",
     "vite": "^5.4.8"
   }
 }
