--- src/app/App.css
@@ -1,122 +1,0 @@
-.app-wrapper {
-  display: grid;
-  grid-template-columns: 18vw auto;
-  grid-template-rows: 150px auto;
-  min-height: 100vh;
-}
-
-.sidebar h1 {
-  font-size:56px;
-  margin-top:10px;    
-}
-
-.header-seperator {
-  grid-row:2/3;
-  grid-column: 2/3;
-
-  border-width: 2.5px;
-  margin:0;
-  width:100%;
-}
-
-.vertical-seperator {
-  grid-row:1/4;
-  grid-column: 2;
-
-  border-width: 2.5px 0 0 2.5px;
-  margin: 0;
-  width:0px;
-}
-
-/* #region sidebar */
-.sidebar {
-  grid-column: 1;
-  grid-row:1/2;
-
-  width: 18vw;
-  min-height: 500px;
-
-  display:flex;
-  flex-direction: column;
-}
-
-.sidebar-divider {
-  width: inherit;
-  border-width: 1.5px;
-  margin:0;
-}
-
-.sidebar-preset {
-  padding: 0 0 5px 10px;
-}
-
-.sidebar-item {
-  border:0;
-  border-radius: 5px;
-
-  margin: 15px 7.5px;
-  padding-left: 10px;
-  height: 2.5em;
-  display: flex;
-  align-items: center;
-
-  color:black;
-  font-size: 24px;
-  transition-duration: 250ms;
-}
-
-.sidebar-item:hover {
-  background-color: var(--dark);
-  cursor: pointer;
-}
-
-.sidebar-item-icon {
-  width: 2em;
-}
-
-/* #endregion */
-
-/* #region header */
-.header {
-  grid-column: 2/3;
-  grid-row: 1/2;
-
-  display:flex;
-  justify-content: space-between;
-  flex: 1;
-  max-height:150px;
-
-  margin:10px 0;
-}
-
-.navbar-item, .navbar-item-today {
-  margin: 20px;
-  padding: 0px 7.5px;
-  
-  flex: 1;
-  transition-duration: 350ms;
-
-  border: 5px solid var(--outline);
-  border-radius: 10px;
-
-  font-size: 24px;
-}
-
-.navbar-item-today {
-  background-color: var(--standard);
-}
-
-.navbar-item:hover,.navbar-item-today:hover {
-  margin: 13px;
-  background-color: var(--bg-highlight);
-}
-
-/* #endregion */
-
-/* #region main content WRITE CSS HERE TASHA */
-.main-content {
-  grid-row: 2;
-  grid-column: 2;
-}
-
-/* #endregion */
