+++ src/app/App.jsx
@@ -1,8 +1,6 @@
-import { useState } from 'react'
-import './App.css'
-import Navbar from './components/navbar/navbar'
-import MainContent from './components/mainContent/mainContent'
-import Sidebar from './components/sidebar/sidebar'
+import Navbar from './components/navbar'
+import MainContent from './components/mainContent'
+import Sidebar from './components/sidebar'
 
 function App() {
