+++ src/app/components/mainContent.jsx
@@ -1,5 +1,4 @@
 function MainContent () {
-
     return <div class="main-content containers">main content</div>
 }
