--- src/app/components/sidebar/sidebar.jsx
@@ -1,13 +1,0 @@
-function Sidebar () {
-    
-    return (
-        <div class="sidebar containers">
-            <h1>To-do List</h1>
-            <div class="sidebar-presets"></div>
-            <hr class="sidebar-divider"></hr>
-            <div class="sidebar-custom"></div>
-        </div>
-    );
-};
-
-export default Sidebar;
