+++ src/stylesheets/_app.scss
@@ -1,0 +1,43 @@
+@forward 'sidebar';
+@forward 'navbar';
+@forward 'variables';
+@use 'variables' as *;
+
+
+.app-wrapper {
+  display: grid;
+  grid-template-columns: 18vw auto;
+  grid-template-rows: 150px auto;
+  min-height: 100vh;
+}
+
+.sidebar h1 {
+  font-size:56px;
+  margin-top:10px;    
+}
+
+.header-seperator {
+  grid-row:2/3;
+  grid-column: 2/3;
+
+  border-width: 2.5px;
+  margin:0;
+  width:100%;
+}
+
+.vertical-seperator {
+  grid-row:1/4;
+  grid-column: 2;
+
+  border-width: 2.5px 0 0 2.5px;
+  margin: 0;
+  width:0px;
+}
+
+/* #region main content WRITE CSS HERE TASHA */
+.main-content {
+  grid-row: 2;
+  grid-column: 2;
+}
+
+/* #endregion */
