+++ src/stylesheets/_navbar.scss
@@ -1,0 +1,35 @@
+@use 'variables' as *;
+
+.header {
+    grid-column: 2/3;
+    grid-row: 1/2;
+  
+    display:flex;
+    justify-content: space-between;
+    flex: 1;
+    max-height:150px;
+  
+    margin:10px 0;
+  
+  .navbar-item, .navbar-item-today {
+    margin: 20px;
+    padding: 0px 7.5px;
+    
+    flex: 1;
+    transition-duration: 350ms;
+  
+    border: 5px solid $outline;
+    border-radius: 10px;
+  
+    font-size: 24px;
+  }
+  
+  .navbar-item-today {
+    background-color: $standard;
+  }
+  
+  .navbar-item:hover,.navbar-item-today:hover {
+    margin: 13px;
+    background-color: $bg-highlight;
+  }
+}
