+++ src/stylesheets/_variables.scss
@@ -1,0 +1,9 @@
+$accent: #E75A5F;  
+$light: #FAFADC;
+$standard: #FEE698;
+$dark: #FDB15D;
+
+// presets
+$bg-color: $light;
+$outline: $accent;
+$bg-highlight: $dark;
