+++ test.rest
@@ -1,0 +1,1 @@
+GET http://localhost:6969/api/doGetDefaultLists
